+++ src/cli/cli.ts
@@ -1,8 +1,10 @@
 #!/usr/bin/env node
 import * as dotenv from 'dotenv';
+
 dotenv.config();
 import arg from 'arg';
 import chalk from 'chalk';
+import { snapshot } from './snapshot';
 import { create } from './create';
 import { run } from './run';
 import { build } from './build';
@@ -10,10 +12,12 @@ import { set } from './set';
 import { test } from './test';
 import { verify } from './verify';
 import { convert } from './convert';
-import {additionalHelpMessages, buildHelpMessage, help} from './help';
+import { additionalHelpMessages, buildHelpMessage, help } from './help';
+import { pack } from "./pack";
 import { InquirerUIProvider } from '../ui/InquirerUIProvider';
 import { argSpec, Runner, RunnerContext } from './Runner';
 import { getConfig } from '../config/utils';
+import { rename } from './rename';
 
 const runners: Record<string, Runner> = {
     create,
@@ -24,6 +28,9 @@ const runners: Record<string, Runner> = {
     help,
     verify,
     convert,
+    rename,
+    pack,
+    snapshot,
 };
 
 async function main() {
@@ -69,7 +76,8 @@ async function main() {
     const runner = effectiveRunners[command];
     if (!runner) {
         console.log(
-            chalk.redBright(`Error: command ${command} not found.`) + `\nRunning ${chalk.cyanBright('blueprint help')}...`
+            chalk.redBright(`Error: command ${command} not found.`) +
+                `\nRunning ${chalk.cyanBright('blueprint help')}...`,
         );
         const helpMessage = buildHelpMessage();
         console.log(helpMessage);