+++ src/cli/constants.ts
@@ -27,9 +27,21 @@ export const templateTypes: { name: string; value: string }[] = [
     },
 ];
 
-export const helpArgs = { '--help': Boolean };
+export const helpArgs = { '--help': Boolean, '-h': '--help' };
 
-const availableCommands = ['create', 'run', 'build', 'set', 'help', 'test', 'verify', 'convert'];
+const availableCommands = [
+    'create',
+    'run',
+    'build',
+    'set',
+    'help',
+    'test',
+    'verify',
+    'convert',
+    'rename',
+    'pack',
+    'snapshot',
+];
 
 export const helpMessages = {
     help: `${chalk.bold('Usage:')} blueprint ${chalk.cyan('help')} [${chalk.yellow('command')}]
@@ -54,7 +66,7 @@ ${chalk.cyan('--type')} <type> - specifies the template type to use when creatin
 ${chalk.bold('List of available types:')}
 ${templateTypes.map((t) => `${chalk.cyan(t.value)} - ${t.name}`).join('\n')}`,
 
-    run: `${chalk.bold('Usage:')} blueprint ${chalk.cyan('run')} ${chalk.yellow('[script name]')} ${chalk.gray('[flags]')}
+    run: `${chalk.bold('Usage:')} blueprint ${chalk.cyan('run')} ${chalk.yellow('[script name]')} ${chalk.gray('[flags]')} ${chalk.gray('[...args]')}
 
 Runs a script from the scripts directory.
 
@@ -67,7 +79,12 @@ ${chalk.cyan('--custom-version')} - API version (v2, v4)
 ${chalk.cyan('--custom-key')} - API key (v2 only)
 ${chalk.cyan('--custom-type')} - network type (custom, mainnet, testnet)
 ${chalk.cyan('--tonconnect')}, ${chalk.cyan('--deeplink')}, ${chalk.cyan('--mnemonic')} - deployer options
-${chalk.cyan('--tonscan')}, ${chalk.cyan('--tonviewer')}, ${chalk.cyan('--toncx')}, ${chalk.cyan('--dton')} - explorer (default: tonviewer)`,
+${chalk.cyan('--tonscan')}, ${chalk.cyan('--tonviewer')}, ${chalk.cyan('--toncx')}, ${chalk.cyan('--dton')} - explorer (default: tonviewer)
+${chalk.gray('[...args]')} (array of strings, optional) - Arguments passed directly to the script.
+                            
+${chalk.bold('Examples:')}
+blueprint run ${chalk.yellow('deployCounter')} ${chalk.cyan('--testnet')} ${chalk.cyan('--tonconnect')}
+blueprint run ${chalk.yellow('incrementCounter')} ${chalk.cyan('--testnet')} ${chalk.cyan('--tonconnect')} ${chalk.gray('0.05 1')}`,
 
     build: `${chalk.bold('Usage:')} blueprint ${chalk.cyan('build')} ${chalk.yellow('[contract name]')} ${chalk.gray('[flags]')}
 
@@ -81,9 +98,15 @@ ${chalk.cyan('--all')} - builds all available contracts.`,
 ${chalk.bold('Available keys:')}
 - ${chalk.cyan('func')} - overrides @ton-community/func-js-bin version.`,
 
-    test: `${chalk.bold('Usage:')} blueprint ${chalk.cyan('test')} [...args]
+    test: `${chalk.bold('Usage:')} blueprint ${chalk.cyan('test')} ${chalk.yellow('[--gas-report|-g ...args]')}
+Runs ${chalk.green('npm test [...args]')}, which by default executes ${chalk.green('jest')}
 
-Runs ${chalk.green('npm test [...args]')}, which by default executes ${chalk.green('jest')}.`,
+${chalk.bold('Options:')}
+  ${chalk.cyan('--gas-report')}, ${chalk.cyan('-g')} - Run tests and compare with the last snapshot's metrics
+
+${chalk.bold('SEE ALSO')}
+  ${chalk.cyan('blueprint snapshot')}
+`,
 
     verify: `${chalk.bold('Usage:')} blueprint ${chalk.cyan('verify')} ${chalk.yellow('[contract name]')} ${chalk.gray('[flags]')}
 
@@ -99,4 +122,21 @@ ${chalk.cyan('--custom-type')} - network type (mainnet, testnet)`,
     convert: `${chalk.bold('Usage:')} blueprint ${chalk.cyan('convert')} ${chalk.yellow('[path to build script]')}
 
 Attempts to convert a legacy bash build script to a Blueprint compile wrapper.`,
+    rename: `${chalk.bold('Usage:')} blueprint ${chalk.cyan('rename')} ${chalk.yellow('[old contract name (PascalCase)]')} ${chalk.yellow('[new contract name (PascalCase)]')}
+
+Renames contract by exact matching in wrappers, scripts, tests and contracts folders.`,
+    pack: `${chalk.bold('Usage:')} blueprint ${chalk.cyan('pack')}
+
+Builds and prepares a publish-ready package of contract wrappers.
+
+${chalk.bold('Flags:')}
+${chalk.cyan('--no-warn')}, ${chalk.cyan('-n')} - ignore warnings about modifying tsconfig.json, package.json, and removing the dist directory.`,
+    snapshot: `${chalk.bold('Usage:')} blueprint ${chalk.cyan('snapshot')} ${chalk.yellow(
+      '[--label=<comment>|-l=<comment>]',
+    )}
+
+Run with gas usage and cells' sizes collected and write a new snapshot
+
+${chalk.bold('SEE ALSO')}
+  ${chalk.cyan('blueprint test --gas-report')}`,
 };