+++ src/config/CustomNetwork.ts
@@ -1,6 +1,6 @@
 export type CustomNetwork = {
     endpoint: string;
-    version?: 'v2' | 'v4' | 'tonapi';
+    version?: 'v2' | 'v4' | 'tonapi' | 'liteclient';
     key?: string;
     type?: 'mainnet' | 'testnet' | 'custom';
 };